+++ gateway/src/main/java/org/acumos/federation/gateway/config/AdapterConfiguration.java
@@ -22,6 +22,7 @@ package org.acumos.federation.gateway.config;
 
 import org.acumos.federation.gateway.common.Clients;
 import org.acumos.federation.gateway.security.AuthenticationConfiguration;
+import org.acumos.federation.gateway.security.FederationMethodSecurityConfiguration;
 import org.acumos.federation.gateway.service.CatalogService;
 import org.acumos.federation.gateway.service.ContentService;
 import org.acumos.federation.gateway.service.LocalWatchService;
@@ -45,7 +46,8 @@ import org.springframework.scheduling.annotation.EnableScheduling;
  */
 @Configuration
 @Import({TaskConfiguration.class,
-				 AuthenticationConfiguration.class})
+				 AuthenticationConfiguration.class,
+				 FederationMethodSecurityConfiguration.class})
 @EnableConfigurationProperties({FederationInterfaceConfiguration.class,
 																LocalInterfaceConfiguration.class,
 																CDMSClientConfiguration.class,